Move keyboard navigation tests to WysiwygComposer-test.tsx
